Refactor to avoid circular dependencies
